moved edit and delete pet buttons to next to their names
