+++ react-app/src/components/PetPage/PetPage.module.css
@@ -32,7 +32,39 @@
   margin-left: 40px;
 }
 
-.pet_name {
+.util_div {
+  width: 200px;
+  display: flex;
+  justify-content: space-around;
+  align-items: center;
+  margin-right: 20px;
+  margin-left: 20px;
+}
+
+.editing_div,
+.deleting_div {
+  display: flex;
+  align-items: center;
+  height: 30px;
+  border: 1px solid rgba(0, 0, 0, 0.3);
+  border-radius: 3px;
+  padding: 5px 10px;
+  background-color: #ff9400;
+  color: white;
+}
+
+.editing_div:hover,
+.deleting_div:hover {
+  text-decoration: none;
+  transition: background-color 0.2s;
+  background-color: rgba(255, 148, 0, 0.5);
+  cursor: pointer;
+}
+
+.name_and_utils {
+  /* width: 300px; */
+  display: flex;
+  justify-content: space-between;
   margin-bottom: 10px;
 }
